+++ bindings/src/method/client.rs
