+++ bindings/src/method/mod.rs
@@ -1,16 +1,16 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod account_method;
-mod client_method;
-mod wallet_method;
+mod account;
+mod client;
+mod wallet;
 
 use std::fmt::{Formatter, Result as FmtResult};
 
-pub use account_method::AccountMethod;
-pub use client_method::ClientMethod;
+pub use account::AccountMethod;
+pub use client::ClientMethod;
 use iota_sdk::client::secret::SecretManagerDto;
-pub use wallet_method::WalletMethod;
+pub use wallet::WalletMethod;
 
 pub(crate) trait OmittedDebug {
     fn omitted_fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
